requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile --output-file='.\requirements.txt' '.\requirements.in'
#
click==8.1.3
    # via -r .\requirements.in
colorama==0.4.6
    # via click
pillow==9.4.0
    # via -r .\requirements.in
